PrintPlayer 개선
